Documents page and topnav right floating socials
I set up the documents page similarly to the forms page and added the function in style.js "documents(doc)" to differentiate between the forms navigation and the docs navigation. I also set up social media pages in the top right. Instagram, Remind, and Github are now linked in the navbar. This will not be accessible in mobile.